requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
pytest==8.0.2
pytest-cov==4.1.0
openai>=1.0.0
SQLAlchemy
pymysql
cryptography